// frontend/src/shared/lib/api/model/activityProjectionResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityProjectionResponseErrorMessage } from './activityProjectionResponseErrorMessage';

export interface ActivityProjectionResponse {
  projection_id: string;
  status: string;
  value_count: number;
  error_message?: ActivityProjectionResponseErrorMessage;
}